+++ packages/block-tools/package.config.ts
@@ -31,4 +31,5 @@ export default defineConfig({
     noImplicitBrowsersList: 'off',
     noImplicitSideEffects: 'error',
   },
+  dts: 'rolldown',
 })