+++ packages/editor/src/editor/PortableTextEditor.tsx
@@ -29,11 +29,11 @@ import type {
 import {createInternalEditor, type InternalEditor} from './create-editor'
 import {EditorActorContext} from './editor-actor-context'
 import type {EditorActor} from './editor-machine'
+import {eventToChange} from './event-to-change'
 import {PortableTextEditorContext} from './hooks/usePortableTextEditor'
 import type {MutationActor} from './mutation-machine'
 import {RelayActorContext} from './relay-actor-context'
 import type {RelayActor} from './relay-machine'
-import {eventToChange} from './route-events-to-changes'
 import type {SyncActor} from './sync-machine'
 
 const debug = debugWithName('component:PortableTextEditor')