+++ packages/editor/tests/event.list-item.add.test.tsx
@@ -1,0 +1,104 @@
+import {describe, expect, test} from 'vitest'
+import {defineSchema} from '../src'
+import {createTestEditor} from '../src/internal-utils/test-editor'
+import {createTestKeyGenerator} from '../src/internal-utils/test-key-generator'
+
+describe('event.list item.add', () => {
+  test('Scenario: Adding initial level', async () => {
+    const keyGenerator = createTestKeyGenerator()
+    const block = {
+      _key: keyGenerator(),
+      _type: 'block',
+      children: [
+        {_key: keyGenerator(), _type: 'span', text: 'Hello', marks: []},
+      ],
+      style: 'normal',
+      markDefs: [],
+    }
+    const {editorRef} = await createTestEditor({
+      schemaDefinition: defineSchema({
+        lists: [{name: 'bullet'}],
+      }),
+      initialValue: [block],
+    })
+
+    editorRef.current?.send({type: 'focus'})
+
+    editorRef.current?.send({
+      type: 'list item.add',
+      listItem: 'bullet',
+    })
+
+    expect(editorRef.current?.getSnapshot().context.value).toEqual([
+      {
+        ...block,
+        level: 1,
+        listItem: 'bullet',
+      },
+    ])
+  })
+
+  test('Scenario: Preserving existing level', async () => {
+    const keyGenerator = createTestKeyGenerator()
+    const block = {
+      _key: keyGenerator(),
+      _type: 'block',
+      children: [
+        {_key: keyGenerator(), _type: 'span', text: 'Hello', marks: []},
+      ],
+      level: 2,
+      listItem: 'bullet',
+      style: 'normal',
+      markDefs: [],
+    }
+    const {editorRef} = await createTestEditor({
+      schemaDefinition: defineSchema({
+        lists: [{name: 'bullet'}, {name: 'number'}],
+      }),
+      initialValue: [block],
+    })
+
+    editorRef.current?.send({type: 'focus'})
+
+    editorRef.current?.send({
+      type: 'list item.add',
+      listItem: 'number',
+    })
+
+    expect(editorRef.current?.getSnapshot().context.value).toEqual([
+      {
+        ...block,
+        level: 2,
+        listItem: 'number',
+      },
+    ])
+  })
+
+  test('Scenario: Aborting if list type is unknown', async () => {
+    const keyGenerator = createTestKeyGenerator()
+    const block = {
+      _key: keyGenerator(),
+      _type: 'block',
+      children: [
+        {_key: keyGenerator(), _type: 'span', text: 'Hello', marks: []},
+      ],
+      style: 'normal',
+      markDefs: [],
+    }
+    const {editorRef} = await createTestEditor({
+      schemaDefinition: defineSchema({
+        lists: [{name: 'bullet'}],
+      }),
+      initialValue: [block],
+    })
+
+    editorRef.current?.send({type: 'focus'})
+
+    editorRef.current?.send({
+      type: 'list item.add',
+      listItem: 'number',
+    })
+
+    expect(editorRef.current?.getSnapshot().context.value).toEqual([block])
+  })
+})